Move some functions to regular lib
