+++ packages/garbo/src/tasks/barfTurn.ts
@@ -68,11 +68,13 @@ import { GarboStrategy, Macro } from "../combat";
 import { globalOptions } from "../config";
 import { wanderer } from "../garboWanderer";
 import {
+  freeFishyAvailable,
   getBestLuckyAdventure,
   howManySausagesCouldIEat,
   kramcoGuaranteed,
   MEAT_TARGET_MULTIPLIER,
   romanticMonsterImpossible,
+  shouldYachtzee,
   sober,
   targetingMeat,
   willDrunkAdventure,
@@ -105,7 +107,6 @@ import { shouldMakeEgg } from "../resources";
 import { lavaDogsAccessible, lavaDogsComplete } from "../resources/doghouse";
 import { hotTubAvailable } from "../resources/clanVIP";
 import { meatMood } from "../mood";
-import { freeFishyAvailable, shouldYachtzee } from "./yachtzee/lib";
 import { yachtzeeTasks } from "./yachtzee";
 
 const digitizedTarget = () =>
